Reverse class/instance declarations of encrypt/decrypt/recrypt
The primary calling case is via the class methods, which, in turn,
create an instance.  However, this instance is unnecessary, leading to
an extra object allocations.  By reversing the class/instance methods,
we eliminate that object.  Additionally, the protected methods were only
protected because of the instance needing to access them, but when
reversed we can now properly make them private.